feat: add Root AI side panel to document page
Extracts the chat experience from /cms/ai into a reusable RootAIChat
component with page and panel variants, then mounts it as a third
mutually-exclusive right-side panel on the document editor. Toggling
via Cmd/Ctrl+I or the new AI button in the doc status bar opens a fresh
chat that knows which document the user is editing, so prompts like
"add an alt text to this image" target the active doc by default.

The server appends an active-document context block to the system
prompt when the client passes a docId, and the existing CMS tool set
(reads, writes with approval, schema lookups) is unchanged.
