+++ packages/root-cms/ui/pages/DocumentPage/DocumentPage.css
@@ -241,3 +241,44 @@
   overflow: hidden;
   background: white;
 }
+
+.DocumentPage__aiDivider {
+  flex: 0 0 3px;
+  width: 3px;
+  cursor: col-resize;
+  position: relative;
+}
+
+.DocumentPage__aiDivider::before {
+  content: '';
+  display: block;
+  position: absolute;
+  top: 0;
+  height: 100%;
+  left: 1px;
+  width: 1px;
+  background-color: var(--color-border);
+}
+
+.DocumentPage__aiDivider::after {
+  content: '';
+  display: block;
+  position: absolute;
+  top: 0;
+  height: 100%;
+  left: 0;
+  width: 100%;
+  background-color: transparent;
+  transition: background-color 0.18s ease;
+}
+
+.DocumentPage__aiDivider:hover::after {
+  background-color: lightblue;
+}
+
+.DocumentPage__ai {
+  flex: 0 0 420px;
+  max-height: calc(100vh - 48px);
+  overflow: hidden;
+  background: white;
+}
